# Backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tumblr Api Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "localhost:8000",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/block/{blog_id}": {
            "post": {
                "tags": [
                    "Block Blogs"
                ],
                "summary": "block blog",
                "description": " Primary blog block another blog",
                "operationId": "followblog",
                "responses": {
                    "200": {
                        "description": "Successful credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "meta": {
                                            "type": "object",
                                            "example": {
                                                "status": "200",
                                                "msg": "ok"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "meta": {
                                            "type": "object",
                                            "example": {
                                                "status": "404",
                                                "msg": "error"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Block Blogs"
                ],
                "summary": "unblock blog",
                "description": " Primary blog unblock another blog",
                "operationId": "unfollowblog",
                "responses": {
                    "200": {
                        "description": "Successful credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "meta": {
                                            "type": "object",
                                            "example": {
                                                "status": "200",
                                                "msg": "ok"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "meta": {
                                            "type": "object",
                                            "example": {
                                                "status": "404",
                                                "msg": "error"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/block": {
            "get": {
                "tags": [
                    "Block Blogs"
                ],
                "summary": "blocked blogs",
                "description": " Primary blog  get all blocked blogs",
                "operationId": "followersblog",
                "responses": {
                    "200": {
                        "description": "Successful credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "meta": {
                                            "type": "object",
                                            "example": {
                                                "status": "200",
                                                "msg": "ok"
                                            }
                                        },
                                        "response": {
                                            "properties": {
                                                "blocked blog": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "blog_avatar": {
                                                                "type": "string",
                                                                "example": "/storage/imgname2.extension"
                                                            },
                                                            "blog_avatar_shape": {
                                                                "type": "string",
                                                                "example": "circle"
                                                            },
                                                            "blog_username": {
                                                                "type": "string",
                                                                "example": "radwa-ahmed213"
                                                            },
                                                            "blog_id": {
                                                                "type": "integer",
                                                                "example": 1032
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "meta": {
                                            "type": "object",
                                            "example": {
                                                "status": "404",
                                                "msg": "error"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/follow_blog/{blog_id}": {
            "post": {
                "tags": [
                    "Follow Blogs"
                ],
                "summary": "follow blog",
                "description": " Primary blog follow another blog",
                "operationId": "followblog",
                "parameters": [
                    {
                        "name": "blog_id",
                        "in": "path",
                        "description": "Blog_id of followed blog ",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful  response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "meta": {
                                            "type": "object",
                                            "example": {
                                                "status": "200",
                                                "msg": "ok"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "meta": {
                                            "type": "object",
                                            "example": {
                                                "status": "404",
                                                "msg": "not found blog"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "meta": {
                                            "type": "object",
                                            "example": {
                                                "status": "404",
                                                "msg": "Unauthorized"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Follow Blogs"
                ],
                "summary": "unfollow blog",
                "description": " Primary blog unfollow another blog",
                "operationId": "unfollowblog",
                "parameters": [
                    {
                        "name": "blog_id",
                        "in": "path",
                        "description": "Blog_id of unfollowed blog ",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "meta": {
                                            "type": "object",
                                            "example": {
                                                "status": "200",
                                                "msg": "ok"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "meta": {
                                            "type": "object",
                                            "example": {
                                                "status": "404",
                                                "msg": "Unauthorized"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "meta": {
                                            "type": "object",
                                            "example": {
                                                "status": "404",
                                                "msg": "Not found"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/followers/{blog_id}": {
            "get": {
                "tags": [
                    "Follow Blogs"
                ],
                "summary": "follower's blog",
                "description": " Primary blog  get all his/her followers",
                "operationId": "followersblog",
                "parameters": [
                    {
                        "name": "blog_id",
                        "in": "path",
                        "description": "Blog_id of current blog  ",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "meta": {
                                            "type": "object",
                                            "example": {
                                                "status": "200",
                                                "msg": "ok"
                                            }
                                        },
                                        "response": {
                                            "properties": {
                                                "followers": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "blog_avatar": {
                                                                "type": "string",
                                                                "example": "/storage/imgname2.extension"
                                                            },
                                                            "blog_avatar_shape": {
                                                                "type": "string",
                                                                "example": "circle"
                                                            },
                                                            "blog_username": {
                                                                "type": "string",
                                                                "example": "radwa-ahmed213"
                                                            },
                                                            "blog_id": {
                                                                "type": "integer",
                                                                "example": 1032
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "meta": {
                                            "type": "object",
                                            "example": {
                                                "status": "404",
                                                "msg": "Unauthorized"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "meta": {
                                            "type": "object",
                                            "example": {
                                                "status": "404",
                                                "msg": "Not found"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
...
    "security": [
        []
    ]
}